Issue # 430 - Add TangleControllerTest class and refactor setup
